Move Source and Role to Researcher
and add unique_id to Protocol
