+++ projects/admin.py
@@ -1,7 +1,5 @@
 from django.contrib import admin
 
-from .models import Source, Project, Role
+from .models import Project
 
-admin.site.register(Source)
 admin.site.register(Project)
-admin.site.register(Role)
