+++ researchers/models.py
@@ -1,7 +1,8 @@
 from django.contrib.auth.models import User
 from django.db import models
 
-from projects.models import Role
+
+from projects.models import Project
 
 
 class University(models.Model):
@@ -12,6 +13,44 @@ class University(models.Model):
 
     def __str__(self):
         return self.name
+
+
+class Role(models.Model):
+    ROLES = (
+        ('owner', 'Owner'),
+        ('contributor', 'Contributor'),
+        ('watcher', 'Watcher'),
+    )
+
+    researcher = models.ForeignKey('Researcher', related_name='roles')
+    project = models.ForeignKey(Project, related_name='roles', null=True, blank=True)
+    protocol = models.ForeignKey('protocols.Protocol', related_name='roles', null=True, blank=True)
+    role = models.CharField(max_length=255, choices=ROLES)
+
+    class Meta:
+        unique_together = (('researcher', 'project'), ('researcher', 'protocol'))
+
+    def __str__(self):
+        role_object_label = ''
+        if self.project:
+            role_object_label = 'project {}'.format(self.project.name)
+        elif self.protocol:
+            role_object_label = 'protocol {}'.format(self.protocol.name)
+        return '{} {} of/to {}'.format(
+            self.researcher.user.username,
+            self.get_role_display(),
+            role_object_label
+        )
+
+    def clean(self):
+        if not(self.project or self.protocol):
+            raise ValidationError('You must choose either project or protocol!')
+        if self.project and self.protocol:
+            raise ValidationError('You cannot select project and protocol for the same role!')
+
+    @classmethod
+    def get_db_roles(cls):
+        return [role[0] for role in cls.ROLES]
 
 
 class Researcher(models.Model):
@@ -44,3 +83,13 @@ class Researcher(models.Model):
             return Role.objects.filter(
                 user=self.user,
                 role__in=roles,).select_related('project', 'protocol')
+
+
+class Source(models.Model):
+    name = models.CharField(max_length=255)
+    researcher = models.ForeignKey(Researcher, related_name='sources')
+    url = models.URLField(max_length=255, null=True, blank=True)
+    isbn = models.CharField(max_length=255, null=True, blank=True)
+
+    def __str__(self):
+        return self.name
